Updated dashboard and retained the real project

+++ Core/Application/DTO/ExpensePlanningDTO.cs
@@ -3,11 +3,8 @@ namespace Application.DTO.ExpensePlanning{
 
         public int BudgetId {get;set;}
         public int ExpenseTypeId {get;set;}
-        public double Amount{get;set;}
+        public decimal Amount{get;set;}
         public string Description {get;set;}
-        // public DateTime CreatedAt {get;set;}
-        // public int UserAdded{get;set;}
-        // public DateTime UpdatedAt{get;set;}
 
     }
 }
